Move compression methods to internal/flight
