+++ internal/flight/compression_method.go
@@ -1,11 +1,12 @@
 // SPDX-FileCopyrightText: 2026 The Pion community <https://pion.ly>
 // SPDX-License-Identifier: MIT
 
-package dtls
+package flight
 
 import "github.com/pion/dtls/v3/pkg/protocol"
 
-func defaultCompressionMethods() []*protocol.CompressionMethod {
+// DefaultCompressionMethods returns the supported compression methods.
+func DefaultCompressionMethods() []*protocol.CompressionMethod {
 	return []*protocol.CompressionMethod{
 		{},
 	}
